feat(chat): move auth logic to the shared layout component
