+++ apps/chat/src/pages/Layout.tsx
@@ -1,0 +1,141 @@
+import { SessionContextValue, signIn, useSession } from 'next-auth/react';
+import React, { useEffect } from 'react';
+
+import { useTranslation } from 'next-i18next';
+import Head from 'next/head';
+
+import { AuthWindowLocationLike } from '@/src/utils/auth/auth-window-location-like';
+import { delay } from '@/src/utils/auth/delay';
+import { timeoutAsync } from '@/src/utils/auth/timeout-async';
+
+import { Translation } from '../types/translation';
+
+import { AuthActions, AuthSelectors } from '../store/auth/auth.reducers';
+import { useAppDispatch, useAppSelector } from '@/src/store/hooks';
+import {
+  SettingsActions,
+  SettingsSelectors,
+} from '@/src/store/settings/settings.reducers';
+import { UIActions } from '@/src/store/ui/ui.reducers';
+
+import { HomeProps } from '.';
+
+export default function Layout({
+  children,
+  pageProps: { initialState },
+}: {
+  children: React.ReactNode;
+  pageProps: HomeProps;
+}) {
+  const session: SessionContextValue<boolean> = useSession();
+
+  const { t } = useTranslation(Translation.Chat);
+
+  const dispatch = useAppDispatch();
+
+  const isOverlay = useAppSelector(SettingsSelectors.selectIsOverlay);
+
+  const shouldLogin = useAppSelector(AuthSelectors.selectIsShouldLogin);
+  const authStatus = useAppSelector(AuthSelectors.selectStatus);
+
+  const isSignInInSameWindow = useAppSelector(
+    SettingsSelectors.selectIsSignInInSameWindow,
+  );
+
+  const shouldOverlayLogin = isOverlay && shouldLogin;
+
+  // EFFECTS  --------------------------------------------
+  useEffect(() => {
+    if (!isOverlay && shouldLogin) {
+      signIn();
+    }
+  }, [isOverlay, shouldLogin]);
+
+  useEffect(() => {
+    dispatch(AuthActions.setSession(session));
+  }, [dispatch, session]);
+
+  // ON LOAD --------------------------------------------
+
+  useEffect(() => {
+    // Hack for ios 100vh issue
+    const handleSetProperVHPoints = () => {
+      document.documentElement.style.setProperty(
+        '--vh',
+        window.innerHeight * 0.01 + 'px',
+      );
+      dispatch(UIActions.resize());
+    };
+    handleSetProperVHPoints();
+    window.addEventListener('resize', handleSetProperVHPoints);
+
+    dispatch(SettingsActions.initApp());
+  }, [dispatch, initialState]);
+
+  const handleOverlayAuth = async () => {
+    const timeout = 30 * 1000;
+    let complete = false;
+    await Promise.race([
+      timeoutAsync(timeout),
+      (async () => {
+        const authWindowLocation = new AuthWindowLocationLike(
+          `api/auth/signin`,
+          isSignInInSameWindow,
+        );
+
+        await authWindowLocation.ready; // ready after redirects
+        const t = Math.max(100, timeout / 1000);
+        // wait for redirection to back
+        while (!complete) {
+          try {
+            if (authWindowLocation.href === window.location.href) {
+              complete = true;
+              authWindowLocation.destroy();
+              break;
+            }
+          } catch {
+            // Do nothing
+          }
+          await delay(t);
+        }
+        window.location.reload();
+
+        return;
+      })(),
+    ]);
+  };
+
+  return (
+    <>
+      <Head>
+        <title className="whitespace-pre">
+          {initialState.settings.appName}
+        </title>
+        <meta name="description" content="ChatGPT but better." />
+        <meta
+          name="viewport"
+          content="height=device-height ,width=device-width, initial-scale=1, user-scalable=no"
+        />
+      </Head>
+      {shouldOverlayLogin ? (
+        <div className="grid size-full min-h-[100px] place-items-center bg-layer-1 text-sm text-primary">
+          <button
+            onClick={handleOverlayAuth}
+            className="button button-secondary"
+            disabled={authStatus === 'loading'}
+          >
+            {t('Login')}
+          </button>
+        </div>
+      ) : (
+        <main
+          // eslint-disable-next-line tailwindcss/enforces-shorthand
+          className="h-screen w-screen flex-col bg-layer-1 text-sm text-primary"
+          id="theme-main"
+        >
+          {children}
+        </main>
+      )}
+    </>
+  );
+}
